+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix to work without a tracing
 - [**BC**] Remove `Logger` type and use `ILogger` from `Microsoft.Extensions.Logging`
 - [**BC**] Add configuration option to set `CommitMessage` *mode*
+- [**BC**] Require qualified access for `GroupId`
 
 ## 16.2.0 - 2021-09-14
 - Update dependencies